Good enough to keep
